pyproject.toml
[project]
name = "fastapi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "uvicorn>=0.41.0",
    "pydantic",
    "sqlalchemy",
    "psycopg2-binary",
    "python-dotenv",
    "passlib",
    "argon2-cffi",
    "python-jose",
    "ultralytics",
]
